requirements.txt
Django==5.2
djangorestframework==3.16.0
mysqlclient==2.2.7
python-decouple==3.8
